+++ Components/WinFormUI/Events/WinFormUIButtonTriggered.cs
@@ -1,13 +1,13 @@
 ﻿using Slipstream.Shared;
 
-namespace Slipstream.Components.UI.Events
+namespace Slipstream.Components.WinFormUI.Events
 {
-    public class UIButtonTriggered : IEvent
+    public class WinFormUIButtonTriggered : IEvent
     {
-        public string EventType => nameof(UIButtonTriggered);
-        
+        public string EventType => nameof(WinFormUIButtonTriggered);
+
         public IEventEnvelope Envelope { get; set; } = new EventEnvelope();
-        
+
         public string Text { get; set; } = "INVALID-NAME";
     }
 }